Split DataTable stories
