+++ src/js/components/DataTable/stories/Clickable.js
@@ -1,0 +1,23 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react';
+
+import { Grommet, Box, DataTable } from 'grommet';
+import { grommet } from 'grommet/themes';
+
+import { columns, DATA } from './data';
+
+const ClickableDataTable = () => (
+  <Grommet theme={grommet}>
+    <Box align="center" pad="large">
+      {/* eslint-disable no-alert */}
+      <DataTable
+        columns={columns}
+        data={DATA}
+        step={10}
+        onClickRow={event => alert(JSON.stringify(event.datum, null, 2))}
+      />
+    </Box>
+  </Grommet>
+);
+
+storiesOf('DataTable', module).add('Clickable', () => <ClickableDataTable />);
