+++ src/js/components/DataTable/stories/Controlled.js
@@ -1,0 +1,65 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react';
+
+import { Grommet, Box, DataTable, CheckBox } from 'grommet';
+import { grommet } from 'grommet/themes';
+
+import { columns, DATA } from './data';
+
+const controlledColumns = columns.map(col => ({ ...col }));
+delete controlledColumns[0].footer;
+delete controlledColumns[3].footer;
+delete controlledColumns[4].footer;
+delete controlledColumns[4].aggregate;
+
+const ControlledDataTable = () => {
+  const [checked, setChecked] = React.useState([]);
+
+  const onCheck = (event, value) => {
+    if (event.target.checked) {
+      setChecked([...checked, value]);
+    } else {
+      setChecked(checked.filter(item => item !== value));
+    }
+  };
+
+  const onCheckAll = event =>
+    setChecked(event.target.checked ? DATA.map(datum => datum.name) : []);
+
+  return (
+    <Grommet theme={grommet}>
+      <Box align="center" pad="medium">
+        <DataTable
+          columns={[
+            {
+              property: 'checkbox',
+              render: datum => (
+                <CheckBox
+                  key={datum.name}
+                  checked={checked.indexOf(datum.name) !== -1}
+                  onChange={e => onCheck(e, datum.name)}
+                />
+              ),
+              header: (
+                <CheckBox
+                  checked={checked.length === DATA.length}
+                  indeterminate={
+                    checked.length > 0 && checked.length < DATA.length
+                  }
+                  onChange={onCheckAll}
+                />
+              ),
+              sortable: false,
+            },
+            ...controlledColumns,
+          ].map(col => ({ ...col }))}
+          data={DATA}
+          sortable
+          size="medium"
+        />
+      </Box>
+    </Grommet>
+  );
+};
+
+storiesOf('DataTable', module).add('Controlled', () => <ControlledDataTable />);
