+++ src/js/components/DataTable/stories/Simple.js
@@ -1,0 +1,17 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react';
+
+import { Grommet, Box, DataTable } from 'grommet';
+import { grommet } from 'grommet/themes';
+
+import { columns, DATA } from './data';
+
+const SimpleDataTable = () => (
+  <Grommet theme={grommet}>
+    <Box align="center" pad="large">
+      <DataTable columns={columns} data={DATA} step={10} />
+    </Box>
+  </Grommet>
+);
+
+storiesOf('DataTable', module).add('Simple', () => <SimpleDataTable />);
